AUDIOX/services/text-extractor/requirements.txt
redis
requests
pymupdf
pdf2image
pytesseract
pillow